disentangle base58 (ocaml) dependency from binable
